% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fcn_add_date_interval}
\alias{fcn_add_date_interval}
\title{Add date interval id to all of the dates}
\usage{
fcn_add_date_interval(df, name_field = "time_period_id")
}
\arguments{
\item{df:}{dataframe with column Date}
}
\description{
Add date interval id to all of the dates
}
