% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-autocorrelation.R
\name{fcn_adj_matrix}
\alias{fcn_adj_matrix}
\title{Generate adjacency matrix for spatial autocorrelation using a secondary grid, and lookup table from primary grid to secondary grid}
\usage{
fcn_adj_matrix(secondary_grid_size = 2000, directions = "queen")
}
\arguments{
\item{secondary_grid_size:}{dimensions of the secondary grid, in m}

\item{directions:}{either "queen" or "rook", for calculating the contiguity weights matrix}
}
\description{
Generate adjacency matrix for spatial autocorrelation using a secondary grid, and lookup table from primary grid to secondary grid
}
