% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-of-area.R
\name{fcn_all_of_area_all}
\alias{fcn_all_of_area_all}
\title{Extract perpendicular distances for all databases}
\usage{
fcn_all_of_area_all()
}
\description{
Extract perpendicular distances for all databases
}
