% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-of-area.R
\name{fcn_all_of_area_sf_all}
\alias{fcn_all_of_area_sf_all}
\title{Extract strip transects with spatial representation in polygons}
\usage{
fcn_all_of_area_sf_all()
}
\description{
Extract strip transects with spatial representation in polygons
}
