% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-of-area.R
\name{fcn_all_of_area_table}
\alias{fcn_all_of_area_table}
\title{Generate tables for koala sightings of perpendicular distances}
\usage{
fcn_all_of_area_table(year)
}
\description{
Generate tables for koala sightings of perpendicular distances
}
