% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-of-area.R
\name{fcn_all_of_area_table_2020}
\alias{fcn_all_of_area_table_2020}
\title{Extract perpendicular distances for koala sightings for 2020 database}
\usage{
fcn_all_of_area_table_2020()
}
\description{
Extract perpendicular distances for koala sightings for 2020 database
}
