% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-of-area.R
\name{fcn_all_of_area_table_integrated}
\alias{fcn_all_of_area_table_integrated}
\title{Extract perpendicular distances for koala sightings for integrated database}
\usage{
fcn_all_of_area_table_integrated()
}
\description{
Extract perpendicular distances for koala sightings for integrated database
}
