% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master-extract.R
\name{fcn_all_tables}
\alias{fcn_all_tables}
\title{Extract all tables}
\usage{
fcn_all_tables(
  table_names = c("line_transect", "perp_distance", "strip_transect", "uaoa")
)
}
\description{
Extract master databases from the scripts
}
