% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master-extract.R
\name{fcn_all_tables_sf}
\alias{fcn_all_tables_sf}
\title{Extract all tables in SF format}
\usage{
fcn_all_tables_sf(table_names = c("line_transect", "strip_transect", "uaoa"))
}
\description{
Extract all tables in SF format
}
