% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-check.R
\name{fcn_check_fractions}
\alias{fcn_check_fractions}
\title{Check if fractions in extracted covariates sum up to approximately 1}
\usage{
fcn_check_fractions(df)
}
\arguments{
\item{df:}{a dataframe with TransectID and fractions columns}
}
\description{
Check if fractions in extracted covariates sum up to approximately 1
}
