% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perp-distance.R
\name{fcn_check_perp_distances_in_line_transect}
\alias{fcn_check_perp_distances_in_line_transect}
\title{Check if all sightings in the line transect is represented in the perpendicular transects table}
\usage{
fcn_check_perp_distances_in_line_transect(line_transect_table, perp_dist_table)
}
\description{
Check if all sightings in the line transect is represented in the perpendicular transects table
}
