% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perp-distance.R
\name{fcn_check_transect_id}
\alias{fcn_check_transect_id}
\title{Check whether transect ID exists in line transect tables}
\usage{
fcn_check_transect_id(db, perp_dist = FALSE)
}
\description{
Check whether transect ID exists in line transect tables
}
