% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-check.R
\name{fcn_check_transect_id_unique}
\alias{fcn_check_transect_id_unique}
\title{Check for duplicate TransectID}
\usage{
fcn_check_transect_id_unique(df)
}
\arguments{
\item{df:}{dataframe with column TransectID}
}
\description{
Functions for checking data correctness
}
