% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_cases.R
\name{fcn_complete_cases_check}
\alias{fcn_complete_cases_check}
\title{Check for complete cases for 2-D and 3-D matrices/ data.frames}
\usage{
fcn_complete_cases_check(mat)
}
\description{
Check for complete cases for 2-D and 3-D matrices/ data.frames
}
