% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcn_line_transect_sf.R
\name{fcn_coord_to_linestring}
\alias{fcn_coord_to_linestring}
\title{Coordinate to linestrings}
\usage{
fcn_coord_to_linestring(
  Start_Eastings,
  Start_Northings,
  End_Eastings,
  End_Northings
)
}
\description{
Coordinate to linestrings
}
