% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_array.R
\name{fcn_cov_array}
\alias{fcn_cov_array}
\title{Extract covariate in array format, with dimensions: GridID, Covariate, Time}
\usage{
fcn_cov_array(
  cov_type = "both",
  dates = NULL,
  time_lag = NULL,
  write_path = NULL
)
}
\arguments{
\item{cov_type:}{covariate types, either "constant", "temporal" or "both". "both" by default}

\item{dates:}{list of date intervals, generated by \code{fcn_date_intervals}. Defaults to start from 1994-10-01 to present in 6-month intervals}

\item{time_lag:}{if cov_type is not "constant", computes additional time-lag variables by number of months specified in an array of months (e.g. time_lag = c(6,12) for computing both 6 month and 12 month time lags). No time lags computed by default (i.e. time_lag = NULL)}
}
\description{
Functions for extracting covariates in a array/ data.table format
}
