% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_extraction.R
\name{fcn_cov_grid_df}
\alias{fcn_cov_grid_df}
\title{Extract covariate in data frame format with GridID}
\usage{
fcn_cov_grid_df(cov = NULL, buffer = c(0, 2500))
}
\arguments{
\item{cov:}{the output of fcn_extract_covariate_grid}
}
\description{
Extract covariate in data frame format with GridID
}
