% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_extraction.R
\name{fcn_covariate_df_match_date}
\alias{fcn_covariate_df_match_date}
\title{Match date for a dataframe with transect Dates and multi-temporal covariate layers}
\usage{
fcn_covariate_df_match_date(df)
}
\description{
Match date for a dataframe with transect Dates and multi-temporal covariate layers
}
