% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_array.R
\name{fcn_covariate_interval_mean}
\alias{fcn_covariate_interval_mean}
\title{Get mean values of each Grid cell given the name of covariate and a single date}
\usage{
fcn_covariate_interval_mean(date, cov_name)
}
\arguments{
\item{date:}{a list object from \code{fcn_date_intervals} containing: interval, start_date, end_date, middle_date}

\item{cov_name:}{name of the covariate (excluding dates of the covariate layer)}
}
\description{
Get mean values of each Grid cell given the name of covariate and a single date
}
