% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_extraction.R
\name{fcn_covariate_match_date}
\alias{fcn_covariate_match_date}
\title{Get covariate value from a spatio-temporal raster}
\usage{
fcn_covariate_match_date(route_table, col_names, method = "bilinear")
}
\description{
Get covariate value from a spatio-temporal raster
}
