% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_extraction.R
\name{fcn_covariate_raster_load}
\alias{fcn_covariate_raster_load}
\title{Load covariate layer to SpatRaster by name}
\usage{
fcn_covariate_raster_load(covariate = "htele")
}
\arguments{
\item{covariate}{name of covariate (string)}
}
\description{
Load covariate layer to SpatRaster by name
}
