% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_extraction.R
\name{fcn_covariate_raster_stack}
\alias{fcn_covariate_raster_stack}
\title{Get all covariates as raster stacks}
\usage{
fcn_covariate_raster_stack(date_interval = NULL, by_date = TRUE)
}
\description{
Get all covariates as raster stacks
}
