% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_extraction.R
\name{fcn_covariate_read_raster}
\alias{fcn_covariate_read_raster}
\title{Read raster}
\usage{
fcn_covariate_read_raster(
  covariate_files,
  covariate_names = NA,
  project = T,
  resample_to_grid = T
)
}
\description{
Read raster
}
