% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_extraction.R
\name{fcn_covariate_temporal_raster_stack}
\alias{fcn_covariate_temporal_raster_stack}
\title{Get temporal covariates as raster stacks}
\usage{
fcn_covariate_temporal_raster_stack(cov_name = NULL, date_interval = NULL)
}
\description{
Get temporal covariates as raster stacks
}
