% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{fcn_date_bilinear_interpolation_date}
\alias{fcn_date_bilinear_interpolation_date}
\title{Convert date difference matrix to sparse matrix of weights of the two closest dates for temporal interpolation}
\usage{
fcn_date_bilinear_interpolation_date(date_diff_matrix)
}
\description{
Convert date difference matrix to sparse matrix of weights of the two closest dates for temporal interpolation
}
