% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{fcn_date_diff_matrix}
\alias{fcn_date_diff_matrix}
\title{Get a matrix of date difference from a vector of reference dates with data}
\usage{
fcn_date_diff_matrix(dates, ref_dates)
}
\description{
Get a matrix of date difference from a vector of reference dates with data
}
