% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_array.R
\name{fcn_date_intervals}
\alias{fcn_date_intervals}
\title{Get dates from start to end in an interval format}
\usage{
fcn_date_intervals(
  start_date = "1994-10-01",
  end_date = NULL,
  interval_months = 6
)
}
\arguments{
\item{start_date:}{start date}

\item{end_date:}{end date (if NULL, taken as the current date)}

\item{interval_months:}{interval in months ("6 months", or "12 months")}
}
\description{
Get dates from start to end in an interval format
}
