% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{fcn_date_weights_nearest_date}
\alias{fcn_date_weights_nearest_date}
\title{Convert date difference matrix to sparse matrix of nearest dates}
\usage{
fcn_date_weights_nearest_date(date_diff_list)
}
\description{
Convert date difference matrix to sparse matrix of nearest dates
}
