% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect.R
\name{fcn_db_connect}
\alias{fcn_db_connect}
\title{Generate a database connection}
\usage{
fcn_db_connect(db_path)
}
\arguments{
\item{db_path}{Path to the database}
}
\description{
Generate a database connection
}
