% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-check.R
\name{fcn_db_to_date}
\alias{fcn_db_to_date}
\title{Convert date in the integrated database to a date object}
\usage{
fcn_db_to_date(df, fmt = "\%d/\%m/\%Y")
}
\arguments{
\item{df}{data frame}
}
\description{
Convert date in the integrated database to a date object
}
