% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_array.R
\name{fcn_extract_cov_date}
\alias{fcn_extract_cov_date}
\title{Extract covariates for a particular date object}
\usage{
fcn_extract_cov_date(d, time_lag = NULLL)
}
\arguments{
\item{d}{date object}
}
\description{
Extract covariates for a particular date object
}
