% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_extraction.R
\name{fcn_extract_covariate_grid}
\alias{fcn_extract_covariate_grid}
\title{Extract all covariate by fishnet grid}
\usage{
fcn_extract_covariate_grid(
  filename_list = NULL,
  fishnet = NULL,
  proportion = FALSE
)
}
\arguments{
\item{filename_list:}{list of covariates by filenames, if unspecified extracts all files}

\item{proportion:}{whether to calculate proportions of categorical variable rather than mode}
}
\description{
Extract all covariate by fishnet grid
}
