% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master-extract.R
\name{fcn_extract_raster_buffer}
\alias{fcn_extract_raster_buffer}
\title{Extract raster after buffering polygons}
\usage{
fcn_extract_raster_buffer(df, fishnet, buffer = 0)
}
\description{
Extract raster after buffering polygons
}
