% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishnet.R
\name{fcn_fishnet}
\alias{fcn_fishnet}
\title{Generate grid fishnet with an input spatial feature class}
\usage{
fcn_fishnet(feature_class)
}
\arguments{
\item{feature_class}{An SF object of the spatial extent of the fishnet}

\item{grid_size}{Grid size in km
Generate a fishnet grid with the input feature_class}
}
\description{
Generate grid fishnet with an input spatial feature class
}
