% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishnet.R
\name{fcn_fishnet_raster}
\alias{fcn_fishnet_raster}
\title{Get fishnet in raster format}
\usage{
fcn_fishnet_raster(feature_class, grid_size = NULL, mask = T)
}
\arguments{
\item{feature_class}{An SF object of the spatial extent of the fishnet}
}
\description{
Get fishnet in raster format
}
