% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_extraction.R
\name{fcn_get_covariate_name}
\alias{fcn_get_covariate_name}
\title{Extract name from string containing dates}
\usage{
fcn_get_covariate_name(input)
}
\description{
Extract name from string containing dates
}
