% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the.R
\name{fcn_get_state}
\alias{fcn_get_state}
\title{Get the whole state, or elements of the state if a second argument is specified}
\usage{
fcn_get_state(elem = NULL)
}
\description{
Get the whole state, or elements of the state if a second argument is specified
}
