% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_extraction.R
\name{fcn_grid_intersect_feature}
\alias{fcn_grid_intersect_feature}
\title{Given a Shapefile (in R sf format), return the GridID intersection with the feature class}
\usage{
fcn_grid_intersect_feature(feature_class, field = NULL, impute_buffer = 0)
}
\arguments{
\item{feature_class:}{a shapefile (in R sf format)}

\item{field:}{field in the feature class to use in the lookup table}

\item{impute_buffer:}{buffer to impute cells without polygon overlap with. Useful for cases where the polygon does not cover the whole study area fully. If 0, there will be no imputation (default), and if >0, then it is the distance of the moving window used for imputation.}
}
\description{
Given a Shapefile (in R sf format), return the GridID intersection with the feature class
}
