% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_extraction.R
\name{fcn_impute_raster}
\alias{fcn_impute_raster}
\title{Impute rasters based on the focal value}
\usage{
fcn_impute_raster(raster_file, name = NA)
}
\description{
Impute rasters based on the focal value
}
