% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_array.R
\name{fcn_impute_temporal_cov}
\alias{fcn_impute_temporal_cov}
\title{Impute missing temporal covariates}
\usage{
fcn_impute_temporal_cov(cov)
}
\arguments{
\item{cov:}{Covariate array}
}
\description{
Impute missing temporal covariates
}
