% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{fcn_interpolation_weight}
\alias{fcn_interpolation_weight}
\title{Given a vector of date differences, find interpolation weight of the n values smallest in the vector}
\usage{
fcn_interpolation_weight(v)
}
\description{
Given a vector of date differences, find interpolation weight of the n values smallest in the vector
}
