% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip-transect.R
\name{fcn_join_koala_survey_sites}
\alias{fcn_join_koala_survey_sites}
\title{Join any table with the KoalaSurveySites to get spatial representation}
\usage{
fcn_join_koala_survey_sites(table, site_id_col = "SiteID")
}
\description{
Join any table with the KoalaSurveySites to get spatial representation
}
