% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master-extract.R
\name{fcn_keep_distinct}
\alias{fcn_keep_distinct}
\title{Remove observations where the observations cannot be identified uniquely, and warn accordingly}
\usage{
fcn_keep_distinct(df, cols = c("SiteNumber", "TransectNumber", "SurveyNumber"))
}
\description{
Remove observations where the observations cannot be identified uniquely, and warn accordingly
}
