% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcn_line_transect_sf.R
\name{fcn_line_transect_sf}
\alias{fcn_line_transect_sf}
\title{Create spatial sf object from start and end point of line transects with eastings and northings}
\usage{
fcn_line_transect_sf(
  lineTransect,
  colnames = list(Start_Eastings = "Start_Eastings", Start_Northings = "Start_Northings",
    End_Eastings = "End_Eastings", End_Northings = "End_Northings"),
  start_point = F
)
}
\arguments{
\item{lineTransect}{A dataframe with columns of Start_Eastings, Start_Northings, End_Eastings, and End_Northings}

\item{colnames}{A list with items Start_Eastings, Start_Northings, End_Eastings, and End_Northings, which are strings corresponding to the column name}

\item{start_point}{A boolean indicating whether the output sf should be just the start points, or the whole table}
}
\description{
Create spatial sf object from start and end point of line transects with eastings and northings
}
