% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_transect.R
\name{fcn_line_transect_sf_all}
\alias{fcn_line_transect_sf_all}
\title{Get all line transects in SF format}
\usage{
fcn_line_transect_sf_all(
  cols = c("TransectID", "SiteID", "Date", "TrSiteID", "Tlength", "Number_Sightings",
    "Number_Observers", "Start_Eastings", "Start_Northings", "End_Eastings",
    "End_Northings", "geometry")
)
}
\description{
Get all line transects in SF format
}
