% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_transect.R
\name{fcn_line_transect_table}
\alias{fcn_line_transect_table}
\title{Extract Line transect tables
Wrapper function for generating line transect tables}
\usage{
fcn_line_transect_table(year)
}
\description{
Extract Line transect tables
Wrapper function for generating line transect tables
}
