% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_extraction.R
\name{fcn_list_covariate_layers_constant}
\alias{fcn_list_covariate_layers_constant}
\title{List constant covariate layers}
\usage{
fcn_list_covariate_layers_constant()
}
\description{
List constant covariate layers
}
