% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_referencing.R
\name{fcn_locate_feature_from_route}
\alias{fcn_locate_feature_from_route}
\title{Route event layer to SF linestrings
Converts Route event layer (from fcn_route_table) to lines}
\usage{
fcn_locate_feature_from_route(route_table, line_transects)
}
\description{
Route event layer to SF linestrings
Converts Route event layer (from fcn_route_table) to lines
}
