% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master-extract.R
\name{fcn_master_covariate_df_match_date}
\alias{fcn_master_covariate_df_match_date}
\title{Update covariate table to match date}
\usage{
fcn_master_covariate_df_match_date(transect_grid, chunk_size = 10000)
}
\description{
Update covariate table to match date
}
