% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_extraction.R
\name{fcn_mixed_extract_raster}
\alias{fcn_mixed_extract_raster}
\title{Extract covariates with mixed multipolygon/ linestring rows}
\usage{
fcn_mixed_extract_raster(input_raster, df)
}
\description{
Extract covariates with mixed multipolygon/ linestring rows
}
