% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perp-distance.R
\name{fcn_perp_distance_all}
\alias{fcn_perp_distance_all}
\title{Extract perpendicular distances for all databases}
\usage{
fcn_perp_distance_all()
}
\description{
Extract perpendicular distances for all databases
}
