% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perp-distance.R
\name{fcn_perp_distance_table}
\alias{fcn_perp_distance_table}
\title{Generate tables for koala sightings of perpendicular distances}
\usage{
fcn_perp_distance_table(year)
}
\description{
Generate tables for koala sightings of perpendicular distances
}
