% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_extract.R
\name{fcn_polygon_extract_raster}
\alias{fcn_polygon_extract_raster}
\title{Extract raster values from polygon}
\usage{
fcn_polygon_extract_raster(input_raster, polygons)
}
\description{
Methods for extracting raster values from polygon
}
