% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_extraction.R
\name{fcn_project_raster}
\alias{fcn_project_raster}
\title{Project raster layer to CRS as specified in the environment}
\usage{
fcn_project_raster(raster)
}
\arguments{
\item{raster}{a \code{SpatRaster} object to be projected}
}
\description{
Project raster layer to CRS as specified in the environment
}
