% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_extraction.R
\name{fcn_read_resample_covariate}
\alias{fcn_read_resample_covariate}
\title{Calculate covariates by resampling by mode}
\usage{
fcn_read_resample_covariate(n, fishnet)
}
\arguments{
\item{n:}{name of covariate}
}
\description{
Calculate covariates by resampling by mode
}
