% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_extraction.R
\name{fcn_read_resample_covariate_proportion}
\alias{fcn_read_resample_covariate_proportion}
\title{Calculate covariates by proportions of categorical values}
\usage{
fcn_read_resample_covariate_proportion(n, fishnet = fishnet)
}
\arguments{
\item{n:}{name of covariate}

\item{fishnet:}{grid raster}
}
\description{
Calculate covariates by proportions of categorical values
}
