% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_referencing.R
\name{fcn_route_table}
\alias{fcn_route_table}
\title{Linear referencing / dynamic segmentation functions
Provides functions to perform linear referencing and dynamic segmentation of line transects.}
\usage{
fcn_route_table(feature_class, line_transect)
}
\arguments{
\item{feature_class}{SF object of the grid}

\item{line_transect}{SF object of the line transect}
}
\description{
Linear referencing / dynamic segmentation functions
Provides functions to perform linear referencing and dynamic segmentation of line transects.
}
