% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_referencing.R
\name{fcn_sample_point_to_route}
\alias{fcn_sample_point_to_route}
\title{Convert covariate sample points to route layer (start and end of covariate value)}
\usage{
fcn_sample_point_to_route(extract_table, names)
}
\arguments{
\item{extract_table}{a sf feature with sampled points, i.e. output of \code{fcn_line_sample_extract_raster}}

\item{names}{names of the covariates in the extract_table}
}
\description{
Convert covariate sample points to route layer (start and end of covariate value)
}
