% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_transect.R
\name{fcn_sf_start_end_coord}
\alias{fcn_sf_start_end_coord}
\title{Get start and end coordinate columns in SF objects with line/ multilinestring objects}
\usage{
fcn_sf_start_end_coord(sf_object)
}
\description{
Get start and end coordinate columns in SF objects with line/ multilinestring objects
}
