% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_array.R
\name{fcn_shift_months}
\alias{fcn_shift_months}
\title{Shift date intervals object by months}
\usage{
fcn_shift_months(date_interval, n_months = -6)
}
\arguments{
\item{date_interval:}{date_interval object from \code{fcn_date_intervals}}

\item{n_months:}{number of months to shift forward or backwards (lag). By default computes a 6 month time lag (-6)}
}
\description{
Shift date intervals object by months
}
