% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect.R
\name{fcn_sql_exec}
\alias{fcn_sql_exec}
\title{SQL extract table by database year and script name}
\usage{
fcn_sql_exec(year, script_name)
}
\description{
SQL extract table by database year and script name
}
