% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip-transect.R
\name{fcn_strip_transect_sf_2015}
\alias{fcn_strip_transect_sf_2015}
\title{Get the table in the 2015-2019 database to its spatial representation}
\usage{
fcn_strip_transect_sf_2015()
}
\description{
Get the table in the 2015-2019 database to its spatial representation
}
