% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip-transect.R
\name{fcn_strip_transect_sf_all}
\alias{fcn_strip_transect_sf_all}
\title{Extract strip transects with spatial representation in polygons}
\usage{
fcn_strip_transect_sf_all()
}
\description{
Extract strip transects with spatial representation in polygons
}
