% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip-transect.R
\name{fcn_strip_transect_table}
\alias{fcn_strip_transect_table}
\title{Generate tables for koala sightings of perpendicular distances}
\usage{
fcn_strip_transect_table(year)
}
\description{
Generate tables for koala sightings of perpendicular distances
}
