% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip-transect.R
\name{fcn_strip_transect_table_2015}
\alias{fcn_strip_transect_table_2015}
\title{Extract strip transects for 2015 database}
\usage{
fcn_strip_transect_table_2015()
}
\description{
Extract strip transects for 2015 database
}
