% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip-transect.R
\name{fcn_strip_transect_table_2020}
\alias{fcn_strip_transect_table_2020}
\title{Extract perpendicular distances for koala sightings for 2020 database}
\usage{
fcn_strip_transect_table_2020()
}
\description{
Extract perpendicular distances for koala sightings for 2020 database
}
