% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip-transect.R
\name{fcn_strip_transect_table_integrated}
\alias{fcn_strip_transect_table_integrated}
\title{Extract strip transects for integrated database}
\usage{
fcn_strip_transect_table_integrated()
}
\description{
Extract strip transects for integrated database
}
