% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_array.R
\name{fcn_temporal_covariate_rds_array}
\alias{fcn_temporal_covariate_rds_array}
\title{Convert RDS of date covariates into a 3-d matrix}
\usage{
fcn_temporal_covariate_rds_array(out_dir = NULL, dates = NULL)
}
\arguments{
\item{out_dir:}{directory where the RDS are stored}
}
\description{
Convert RDS of date covariates into a 3-d matrix
}
