% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master-extract.R
\name{fcn_transect_grid_covariates}
\alias{fcn_transect_grid_covariates}
\title{Join transect grid with covariate based on GridID}
\usage{
fcn_transect_grid_covariates(transect_grid, cov = NULL)
}
\description{
Join transect grid with covariate based on GridID
}
