% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEQKoalaDataPipeline-package.R
\docType{package}
\name{SEQKoalaDataPipeline-package}
\alias{SEQKoalaDataPipeline}
\alias{SEQKoalaDataPipeline-package}
\title{SEQKoalaDataPipeline: South East Queensland Koala Monitoring Data Pipeline}
\description{
Data processing pipeline for koala monitoring data analysis project at South East Queensland.
}
\author{
\strong{Maintainer}: Frankie Cho \email{f.cho@uq.edu.au}

Authors:
\itemize{
  \item Jonathan Rhodes \email{j.rhodes@uq.edu.au}
}

}
\keyword{internal}
