% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the.R
\docType{data}
\name{the}
\alias{the}
\title{Sets environment for storing state and functions for modifying the state}
\format{
An object of class \code{environment} of length 15.
}
\usage{
the
}
\description{
Sets environment for storing state and functions for modifying the state
}
\keyword{datasets}
