% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master-extract.R
\name{fcn_all_transect_grid_fractions}
\alias{fcn_all_transect_grid_fractions}
\title{Extract transect information with grid fractions}
\usage{
fcn_all_transect_grid_fractions(buffer = c(0))
}
\arguments{
\item{buffer:}{a vector of polygon buffer sizes to extract,}
}
\description{
Extract transect information with grid fractions
}
